refactor: PostEntryCommandクラスを新規作成しEntryCommandsから分離
